requirements.txt
vectorbt
numpy
TA-Lib
pandas
alpaca-py
alpaca-trade-api